Remove most uses of Organisation.redirector_abbr
Not all organisations will have passed through redirector, so we need to be
able to cope without it.

The redirector import still needs it, so it remains there.

Additionally, its used in the Google Analytics ingestion code, and I can't see
an easy alternative to redirector_abbr for that use case, other than database
record IDs.